+++ src/frames/frames.ts
@@ -1,0 +1,29 @@
+import { APIC } from './frame-apic'
+import { CHAP } from './frame-chap'
+import { COMM } from './frame-comm'
+import { COMR } from './frame-comr'
+import { CTOC } from './frame-ctoc'
+import { ETCO } from './frame-etco'
+import { POPM } from './frame-popm'
+import { PRIV } from './frame-priv'
+import { SYLT } from './frame-sylt'
+import { TXXX } from './frame-txxx'
+import { UFID } from './frame-ufid'
+import { USLT } from './frame-uslt'
+import { WXXX } from './frame-wxxx'
+
+export const Frames = {
+    APIC,
+    CHAP,
+    COMM,
+    COMR,
+    CTOC,
+    ETCO,
+    POPM,
+    PRIV,
+    SYLT,
+    TXXX,
+    UFID,
+    USLT,
+    WXXX
+}
