+++ src/frames/type.ts
@@ -1,0 +1,4 @@
+
+// TODO: Fix with better types.
+// eslint-disable-next-line
+export type Data = any
